candlestick-provider/src/main/java/com/tr/candlestickprovider/config/CandlestickConfig.java
package com.tr.candlestickprovider.config;

import org.springframework.boot.context.properties.ConfigurationProperties;
import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.context.annotation.Configuration;
import org.springframework.stereotype.Component;

@Configuration
@ConfigurationProperties("candlestick")
public class CandlestickConfig {
    private CandlestickConfig.Partner partner = new CandlestickConfig.Partner();
    public static class Partner {
        private String url;

        public String getUrl() {
            return url;
        }

        public void setUrl(String url) {
            this.url = url;
        }
    }

    public Partner getPartner() {
        return partner;
    }

    public void setPartner(Partner partner) {
        this.partner = partner;
    }
}
